+++ api/trading_test.go
@@ -194,7 +194,7 @@ func getTestGRPCServer(
 	candleService := candles.NewService(logger, conf.Candles, candleStore)
 	marketDataStore := storage.NewMarketData(logger, conf.Storage)
 
-	marketDepth := subscribers.NewMarketDepthBuilder(ctx, logger, true)
+	marketDepth := subscribers.NewMarketDepthBuilder(ctx, logger, nil, false, true)
 	if marketDepth == nil {
 		return
 	}