+++ sqlstore/orders_test.go
@@ -13,18 +13,20 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func addTestOrder(t *testing.T, os *sqlstore.Orders, block entities.Block, party entities.Party, market entities.Market, reference string) entities.Order {
+func addTestOrder(t *testing.T, os *sqlstore.Orders, id entities.OrderID, block entities.Block, party entities.Party, market entities.Market, reference string,
+	side types.Side, timeInForce types.OrderTimeInForce, orderType types.OrderType, status types.OrderStatus,
+	price, size, remaining int64, seqNum uint64) entities.Order {
 	order := entities.Order{
-		ID:              entities.NewOrderID(generateID()),
+		ID:              id,
 		MarketID:        market.ID,
 		PartyID:         party.ID,
-		Side:            types.SideBuy,
-		Price:           10,
-		Size:            100,
-		Remaining:       60,
-		TimeInForce:     types.OrderTimeInForceGTC,
-		Type:            types.OrderTypeLimit,
-		Status:          types.OrderStatusActive,
+		Side:            side,
+		Price:           price,
+		Size:            size,
+		Remaining:       remaining,
+		TimeInForce:     timeInForce,
+		Type:            orderType,
+		Status:          status,
 		Reference:       reference,
 		Version:         1,
 		PeggedOffset:    0,
@@ -33,6 +35,7 @@ func addTestOrder(t *testing.T, os *sqlstore.Orders, block entities.Block, party
 		UpdatedAt:       time.Now().Add(5 * time.Second).Truncate(time.Microsecond),
 		ExpiresAt:       time.Now().Add(10 * time.Second).Truncate(time.Microsecond),
 		VegaTime:        block.VegaTime,
+		SeqNum:          seqNum,
 	}
 
 	err := os.Add(context.Background(), order)
@@ -73,7 +76,21 @@ func TestOrders(t *testing.T) {
 	updatedOrders := []entities.Order{}
 	numOrdersUpdatedInDifferentBlock := 0
 	for i := 0; i < numTestOrders; i++ {
-		order := addTestOrder(t, os, block, parties[i%3], markets[i%2], fmt.Sprintf("my_reference_%d", i))
+		order := addTestOrder(t, os,
+			entities.NewOrderID(generateID()),
+			block,
+			parties[i%3],
+			markets[i%2],
+			fmt.Sprintf("my_reference_%d", i),
+			types.SideBuy,
+			types.OrderTimeInForceGTC,
+			types.OrderTypeLimit,
+			types.OrderStatusActive,
+			10,
+			100,
+			60,
+			uint64(i),
+		)
 		orders = append(orders, order)
 
 		// Don't update 1/4 of the orders
@@ -185,4 +202,215 @@ func reverseOrderSlice(input []entities.Order) (output []entities.Order) {
 		output = append(output, input[i])
 	}
 	return output
+}
+
+func generateTestBlocks(t *testing.T, numBlocks int, bs *sqlstore.Blocks) []entities.Block {
+	t.Helper()
+	blocks := make([]entities.Block, numBlocks, numBlocks)
+	for i := 0; i < numBlocks; i++ {
+		blocks[i] = addTestBlock(t, bs)
+		time.Sleep(time.Millisecond)
+	}
+	return blocks
+}
+
+func generateParties(t *testing.T, numParties int, block entities.Block, ps *sqlstore.Parties) []entities.Party {
+	t.Helper()
+	parties := make([]entities.Party, numParties, numParties)
+	for i := 0; i < numParties; i++ {
+		parties[i] = addTestParty(t, ps, block)
+	}
+	return parties
+}
+
+func addTestMarket(t *testing.T, ms *sqlstore.Markets, block entities.Block) entities.Market {
+	market := entities.Market{
+		ID:       entities.NewMarketID(generateID()),
+		VegaTime: block.VegaTime,
+	}
+
+	err := ms.Upsert(&market)
+	require.NoError(t, err)
+	return market
+}
+
+func generateMarkets(t *testing.T, numMarkets int, block entities.Block, ms *sqlstore.Markets) []entities.Market {
+	t.Helper()
+	markets := make([]entities.Market, numMarkets, numMarkets)
+	for i := 0; i < numMarkets; i++ {
+		markets[i] = addTestMarket(t, ms, block)
+	}
+	return markets
+}
+
+func generateOrderIDs(t *testing.T, numIDs int) []entities.OrderID {
+	t.Helper()
+	orderIDs := make([]entities.OrderID, numIDs)
+	for i := 0; i < numIDs; i++ {
+		orderIDs[i] = entities.NewOrderID(generateID())
+		time.Sleep(time.Millisecond)
+	}
+	return orderIDs
+}
+
+func generateTestOrders(t *testing.T, blocks []entities.Block, parties []entities.Party,
+	markets []entities.Market, orderIDs []entities.OrderID, os *sqlstore.Orders,
+) []entities.Order {
+	// define the orders we're going to insert
+	testOrders := []struct {
+		id          entities.OrderID
+		block       entities.Block
+		party       entities.Party
+		market      entities.Market
+		side        types.Side
+		price       int64
+		size        int64
+		remaining   int64
+		timeInForce types.OrderTimeInForce
+		orderType   types.OrderType
+		status      types.OrderStatus
+	}{
+		{
+			id:          orderIDs[0],
+			block:       blocks[0],
+			party:       parties[0],
+			market:      markets[0],
+			side:        types.SideBuy,
+			price:       100,
+			size:        1000,
+			remaining:   1000,
+			timeInForce: types.OrderTimeInForceGTC,
+			orderType:   types.OrderTypeLimit,
+			status:      types.OrderStatusActive,
+		},
+		{
+			id:          orderIDs[1],
+			block:       blocks[0],
+			party:       parties[1],
+			market:      markets[0],
+			side:        types.SideBuy,
+			price:       101,
+			size:        2000,
+			remaining:   2000,
+			timeInForce: types.OrderTimeInForceGTC,
+			orderType:   types.OrderTypeLimit,
+			status:      types.OrderStatusActive,
+		},
+		{
+			id:          orderIDs[2],
+			block:       blocks[0],
+			party:       parties[2],
+			market:      markets[0],
+			side:        types.SideSell,
+			price:       105,
+			size:        1500,
+			remaining:   1500,
+			timeInForce: types.OrderTimeInForceGTC,
+			orderType:   types.OrderTypeLimit,
+			status:      types.OrderStatusActive,
+		},
+		{
+			id:          orderIDs[3],
+			block:       blocks[0],
+			party:       parties[3],
+			market:      markets[0],
+			side:        types.SideSell,
+			price:       105,
+			size:        800,
+			remaining:   8500,
+			timeInForce: types.OrderTimeInForceGTC,
+			orderType:   types.OrderTypeLimit,
+			status:      types.OrderStatusActive,
+		},
+		{
+			id:          orderIDs[4],
+			block:       blocks[0],
+			party:       parties[0],
+			market:      markets[1],
+			side:        types.SideBuy,
+			price:       1000,
+			size:        10000,
+			remaining:   10000,
+			timeInForce: types.OrderTimeInForceGTC,
+			orderType:   types.OrderTypeLimit,
+			status:      types.OrderStatusActive,
+		},
+		{
+			id:          orderIDs[5],
+			block:       blocks[1],
+			party:       parties[2],
+			market:      markets[1],
+			side:        types.SideSell,
+			price:       1005,
+			size:        15000,
+			remaining:   15000,
+			timeInForce: types.OrderTimeInForceGTC,
+			orderType:   types.OrderTypeLimit,
+			status:      types.OrderStatusActive,
+		},
+		{
+			id:          orderIDs[6],
+			block:       blocks[1],
+			party:       parties[3],
+			market:      markets[2],
+			side:        types.SideSell,
+			price:       1005,
+			size:        15000,
+			remaining:   15000,
+			timeInForce: types.OrderTimeInForceFOK,
+			orderType:   types.OrderTypeMarket,
+			status:      types.OrderStatusActive,
+		},
+		{
+			id:          orderIDs[3],
+			block:       blocks[2],
+			party:       parties[3],
+			market:      markets[0],
+			side:        types.SideSell,
+			price:       1005,
+			size:        15000,
+			remaining:   15000,
+			timeInForce: types.OrderTimeInForceGTC,
+			orderType:   types.OrderTypeLimit,
+			status:      types.OrderStatusCancelled,
+		},
+	}
+
+	orders := make([]entities.Order, len(testOrders))
+
+	for i, to := range testOrders {
+		ref := fmt.Sprintf("reference-%d", i)
+		orders[i] = addTestOrder(t, os, to.id, to.block, to.party, to.market, ref, to.side,
+			to.timeInForce, to.orderType, to.status, to.price, to.size, to.remaining, uint64(i))
+	}
+
+	return orders
+}
+
+func TestOrders_GetLiveOrders(t *testing.T) {
+	defer testStore.DeleteEverything()
+
+	bs := sqlstore.NewBlocks(testStore)
+	ps := sqlstore.NewParties(testStore)
+	ms := sqlstore.NewMarkets(testStore)
+	os := sqlstore.NewOrders(testStore)
+
+	t.Logf("test store port: %d", testDBPort)
+
+	// set up the blocks, parties and markets we need to generate the orders
+	blocks := generateTestBlocks(t, 3, bs)
+	parties := generateParties(t, 5, blocks[0], ps)
+	markets := generateMarkets(t, 3, blocks[0], ms)
+	orderIDs := generateOrderIDs(t, 8)
+	testOrders := generateTestOrders(t, blocks, parties, markets, orderIDs, os)
+
+	// Make sure we flush the batcher and write the orders to the database
+	err := os.Flush(context.Background())
+	require.NoError(t, err)
+
+	want := append(testOrders[:3], testOrders[4:6]...)
+	got, err := os.GetLiveOrders(context.Background())
+	require.NoError(t, err)
+	assert.Equal(t, 5, len(got))
+	assert.ElementsMatch(t, want, got)
 }