+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [](https://github.com/vegaprotocol/data-node/pull/xxx) -
 
 ### 🛠  Improvements
+- [414](https://github.com/vegaprotocol/data-node/issues/414) - Migrate market depth to retrieve data from `Postgres`
 - [495](https://github.com/vegaprotocol/data-node/pull/495) - Remove deprecated `PositionState` event handling, general fixes to `SettlePosition` event handling
 - [](https://github.com/vegaprotocol/data-node/pull/xxx) -